+++ flexmeasures/data/models/planning/charging_station.py
@@ -1,7 +1,7 @@
 from typing import Optional, Union
 from datetime import datetime, timedelta
 
-from pandas import Series, Timestamp
+import pandas as pd
 
 from flexmeasures.data.models.time_series import Sensor
 from flexmeasures.data.models.planning.solver import device_scheduler
@@ -20,12 +20,14 @@ def schedule_charging_station(
     end: datetime,
     resolution: timedelta,
     soc_at_start: float,
-    soc_targets: Series,
+    soc_targets: pd.Series,
     soc_min: Optional[float] = None,
     soc_max: Optional[float] = None,
     roundtrip_efficiency: Optional[float] = None,
     prefer_charging_sooner: bool = True,
-) -> Union[Series, None]:
+    price_sensor: Optional[Sensor] = None,
+    round_to_decimals: Optional[int] = 6,
+) -> Union[pd.Series, None]:
     """Schedule a charging station asset based directly on the latest beliefs regarding market prices within the specified time
     window.
     For the resulting consumption schedule, consumption is defined as positive values.
@@ -52,12 +54,16 @@ def schedule_charging_station(
 
     # Check for known prices or price forecasts, trimming planning window accordingly
     prices, (start, end) = get_prices(
-        sensor, (start, end), resolution, allow_trimmed_query_window=True
+        (start, end),
+        resolution,
+        price_sensor=price_sensor,
+        sensor=sensor,
+        allow_trimmed_query_window=True,
     )
     # soc targets are at the end of each time slot, while prices are indexed by the start of each time slot
     soc_targets = soc_targets.tz_convert("UTC")
-    start = Timestamp(start).tz_convert("UTC")
-    end = Timestamp(end).tz_convert("UTC")
+    start = pd.Timestamp(start).tz_convert("UTC")
+    end = pd.Timestamp(end).tz_convert("UTC")
     soc_targets = soc_targets[start + resolution : end]
 
     # Add tiny price slope to prefer charging now rather than later, and discharging later rather than now.
@@ -133,5 +139,9 @@ def schedule_charging_station(
         )
     else:
         charging_station_schedule = ems_schedule[0]
+
+    # Round schedule
+    if round_to_decimals:
+        charging_station_schedule = charging_station_schedule.round(round_to_decimals)
 
     return charging_station_schedule